Introduce ResolutionParameters
ResolutionParameters are a mostly-immutable but completely thread-safe set of properties used as input
to a dependency resolution. The goal is to have these parameters be the complete set of project-scoped
state required to perform a dependency resolution. Furthermore, this state should remain safe to interact
with without a project lock, meaning once we migrate all project-derived parameters to this class
we can perform dependency resolution without acquiring a project lock.